present user profile ui
