+++ wire-ios/Wire-iOS/Sources/UserInterface/MainController/ZClientViewController+StartUIDelegate.swift
@@ -16,18 +16,19 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import UIKit
 import WireDataModel
-import WireSyncEngine
+// import WireMainNavigationUI
 
 extension ZClientViewController: StartUIDelegate {
 
-    func startUI(_ startUI: StartUIViewController, didSelect user: UserType) {
-        guard let userID = user.qualifiedID else { return }
-
-        let conversation = user.oneToOneConversation
-
+    @MainActor
+    func startUIViewController(_ viewController: StartUIViewController, didSelect user: any UserType) {
         Task {
+            guard let userID = user.qualifiedID else { return }
+
+            let userSession = viewController.userSession
+            let conversation = user.oneToOneConversation
+
             do {
                 let isReady = try await userSession.checkOneOnOneConversationIsReady.invoke(userID: userID)
 
@@ -36,14 +37,14 @@ extension ZClientViewController: StartUIDelegate {
                     // If the conversation exists, and is established (in case of mls),
                     // then we open the conversation
                     guard let conversation else { return }
-                    await openConversation(conversation)
+                    await mainCoordinator.showConversation(conversation: conversation, message: nil)
 
                 } else {
 
                     // If the conversation should be using mls and is not established,
                     // or does not exits, then we open the user profile to let the user
                     // create the conversation
-                    await openUserProfile(user)
+                    await showUserProfile(user: user)
 
                 }
             } catch {
@@ -52,31 +53,10 @@ extension ZClientViewController: StartUIDelegate {
         }
     }
 
-    func startUI(_ startUI: StartUIViewController, didSelect conversation: ZMConversation) {
-        startUI.dismissIfNeeded(animated: true) {
-            ZClientViewController.shared?.select(conversation: conversation, focusOnView: true, animated: true)
+    @MainActor
+    func startUIViewController(_ viewController: StartUIViewController, didSelect conversation: ZMConversation) {
+        Task {
+            await mainCoordinator.showConversation(conversation: conversation, message: nil)
         }
-    }
-
-    @MainActor
-    func openConversation(_ conversation: ZMConversation) {
-        ZClientViewController.shared?.select(conversation: conversation, focusOnView: true, animated: true)
-    }
-
-    @MainActor
-    func openUserProfile(_ user: UserType) {
-        let profileViewController = ProfileViewController(
-            user: user,
-            viewer: selfUserLegalHoldSubject,
-            context: .profileViewer,
-            userSession: userSession,
-            mainCoordinator: mainCoordinator
-        )
-        profileViewController.delegate = self
-
-        let navigationController = profileViewController.wrapInNavigationController()
-        navigationController.modalPresentationStyle = .formSheet
-
-        ZClientViewController.shared?.present(navigationController, animated: true)
     }
 }
