+++ wire-ios/Wire-iOS/Sources/UserInterface/StartUI/Common/ProfilePresenter+ProfileViewControllerDelegate.swift
@@ -21,6 +21,7 @@ import WireDataModel
 extension ProfilePresenter: ProfileViewControllerDelegate {
 
     func profileViewController(_ controller: ProfileViewController?, wantsToNavigateTo conversation: ZMConversation) {
+        // TODO: check diff
         guard let controller else { return }
 
         Task {
